+++ src/main/kotlin/com/wafflestudio/team2server/area/repository/AreaAdjRepository.kt
@@ -3,6 +3,6 @@ package com.wafflestudio.team2server.area.repository
 import com.wafflestudio.team2server.area.model.AreaAdjEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface AreaAdjRepository : JpaRepository<AreaAdjEntity, Long> {
-	fun getAreaAdjEntityByAreaIdAndDistanceIsLessThanEqual(id: Long, distance: Int): List<AreaAdjEntity>
+interface AreaAdjRepository : JpaRepository<AreaAdjEntity, Int> {
+	fun getAreaAdjEntityByAreaIdAndDistanceIsLessThanEqual(id: Int, distance: Int): List<AreaAdjEntity>
 }
